Upgrade checkstyle
I upgraded claimstore checkstyle to remove the hack that we used to fail on
warnings as there is now a property for it, also updated rules to
closely match what we have in the claimstore, happy to undo them if you
don't see any value.

Rules such as no unused imports and hide utility class constructors have
been added.