Moved sample functional and smoke tests to demo directory
Added "demo" directory to functionalTest and smokeTest directory structures so that they match the sample test class package.
Added commands to bin/init.sh to move the sample functional and smoke tests to the directory name provided by package parameter.  This ensures that the directory structure matches the package structure set when the script replaces instances of the "demo" package.