Fix two text replacements in init script
There are two issues with the text replacements made by the init.sh script.  These are:
1.) The packagesToScan property in application.yaml is changed from "uk.gov.hmcts.reform.demo.controllers" to "uk.gov.hmcts.reform.$component_name.controllers".  It should be changed to "uk.gov.hmcts.reform.$package.controllers" to match the directory rename carried out later in the script.
2.) The chart name property in Chart.yaml is changed from "rpe-spring-boot-template" to "$product_name-$product_name-$component_name" by a combination of the maintainer name and slug replacements.  It should be changed to "$product_name-$component_name".

To fix the first issue the changes to the application.yaml file have been moved to a separate section.  This replaces all instances of "rpe" with $product_name and then all instances of "reform.demo" with reform.$package".

To fix the second issue a new line of code has been added to the section that processes the Chart.yaml file.  This removes the "rpe-" prefix from the chart name value leaving just "spring-boot-template".  The remaining "spring-boot-template" part is replaced with "$product_name-$component_name" when the slug replacement is carried out later in the script.